+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/AgentSdJwtInteropTest.kt
@@ -78,7 +78,6 @@ class AgentSdJwtInteropTest : FreeSpec({
             Holder.StoreCredentialInput.SdJwt(input, EuPidScheme)
         ).getOrThrow()
 
-        stored.status shouldBe null
         val entry = stored.storeEntry
         entry.shouldBeInstanceOf<SubjectCredentialStore.StoreEntry.SdJwt>()
         entry.disclosures.size shouldBe 9