+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/CredentialTimelinessValidator.kt
@@ -1,12 +1,9 @@
 package at.asitplus.wallet.lib.agent.validation
 
-import at.asitplus.KmmResult
 import at.asitplus.wallet.lib.agent.SubjectCredentialStore
 import at.asitplus.wallet.lib.agent.validation.mdoc.MdocTimelinessValidator
 import at.asitplus.wallet.lib.agent.validation.sdJwt.SdJwtTimelinessValidator
 import at.asitplus.wallet.lib.agent.validation.vcJws.VcJwsTimelinessValidator
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
-import io.github.aakira.napier.Napier
 import kotlinx.datetime.Clock
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
@@ -14,9 +11,6 @@ import kotlin.time.Duration.Companion.seconds
 data class CredentialTimelinessValidator(
     private val timeLeeway: Duration = 300.seconds,
     private val clock: Clock = Clock.System,
-    private val tokenStatusResolver: TokenStatusResolver = TokenStatusResolver {
-        KmmResult.success(TokenStatus.Valid)
-    },
     val vcJwsTimelinessValidator: VcJwsTimelinessValidator = VcJwsTimelinessValidator(
         timeLeeway = timeLeeway,
         clock = clock
@@ -30,33 +24,28 @@ data class CredentialTimelinessValidator(
         clock = clock
     ),
 ) {
-    suspend operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry) = CredentialTimelinessValidationSummary(
-        tokenStatus = when (storeEntry) {
-            is SubjectCredentialStore.StoreEntry.Iso -> storeEntry.issuerSigned.issuerAuth.payload?.status
-            is SubjectCredentialStore.StoreEntry.SdJwt -> storeEntry.sdJwt.credentialStatus
-            is SubjectCredentialStore.StoreEntry.Vc -> storeEntry.vc.vc.credentialStatus
-        }?.let {
-            tokenStatusResolver(it)
-        },
-        timelinessValidationSummaryDetails = when (storeEntry) {
-            is SubjectCredentialStore.StoreEntry.Iso -> CredentialTimelinessValidationSummaryDetails.MdocCredential(
-                storeEntry = storeEntry,
-                summary = mdocTimelinessValidator.invoke(storeEntry.issuerSigned),
+    operator fun invoke(storeEntry: SubjectCredentialStore.StoreEntry) = when (storeEntry) {
+        is SubjectCredentialStore.StoreEntry.Iso -> invoke(CredentialWrapper.Mdoc(storeEntry.issuerSigned))
+        is SubjectCredentialStore.StoreEntry.SdJwt -> invoke(CredentialWrapper.SdJwt(storeEntry.sdJwt))
+        is SubjectCredentialStore.StoreEntry.Vc -> invoke(CredentialWrapper.VcJws(storeEntry.vc))
+    }
+
+    operator fun invoke(credentialWrapper: CredentialWrapper): CredentialTimelinessValidationSummary {
+        return when (credentialWrapper) {
+            is CredentialWrapper.Mdoc -> CredentialTimelinessValidationSummary.Mdoc(
+                credential = credentialWrapper,
+                mdocTimelinessValidator.invoke(credentialWrapper.issuerSigned),
             )
 
-            is SubjectCredentialStore.StoreEntry.SdJwt -> CredentialTimelinessValidationSummaryDetails.SdJwtCredential(
-                storeEntry = storeEntry,
-                sdJwtTimelinessValidator(storeEntry.sdJwt),
+            is CredentialWrapper.SdJwt -> CredentialTimelinessValidationSummary.SdJwt(
+                credential = credentialWrapper,
+                sdJwtTimelinessValidator(credentialWrapper.sdJwt),
             )
 
-            is SubjectCredentialStore.StoreEntry.Vc -> CredentialTimelinessValidationSummaryDetails.VerifiableCredential(
-                storeEntry = storeEntry,
-                vcJwsTimelinessValidator(vcJws = storeEntry.vc),
+            is CredentialWrapper.VcJws -> CredentialTimelinessValidationSummary.VcJws(
+                credential = credentialWrapper,
+                vcJwsTimelinessValidator(vcJws = credentialWrapper.verifiableCredentialJws),
             )
-        }
-    ).also {
-        if (it.isSuccess) {
-            Napier.d("VC is timely")
         }
     }
 }