+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/validation/sdJwt/SdJwtCredentialPayloadValidationSummary.kt
@@ -6,4 +6,7 @@ import at.asitplus.wallet.lib.data.VerifiableCredentialSdJwt
 data class SdJwtCredentialPayloadValidationSummary(
     val verifiableCredentialSdJwt: VerifiableCredentialSdJwt,
     val subjectMatchingResult: SubjectMatchingResult?,
-)
+) {
+    val isSuccess: Boolean
+        get() = subjectMatchingResult?.isSuccess != false
+}